Split policies/timeouts.js into domain submodules + add node test harness (#1037)
Splits monolithic policies/timeouts.js (1447 lines) into 10 domain submodules under policies/timeouts/ (idle-kill, active-monitor, long-turn-monitor, card-timeouts, dispatch-maintenance, orphan-dispatch, reconciliation, review-auto-accept, review-timeouts, workspace-branch-guard). Shared helpers extracted to policies/lib/timeouts-helpers.js. Node-based test harness under policies/__tests__/support/ with mocked adk globals covers all submodules (11 tests). Loader extended to accept modular policies. CI workflows updated to run node tests alongside cargo tests. Runtime policy behavior preserved.
