+++ .github/workflows/ci-main.yml
@@ -36,6 +36,9 @@ jobs:
               - 'policies/auto-queue.js'
               - 'policies/kanban-rules.js'
               - 'policies/timeouts.js'
+              - 'policies/timeouts/**'
+              - 'policies/lib/**'
+              - 'policies/__tests__/**'
               - 'src/db/**'
               - 'src/dispatch/**'
               - 'src/engine/**'
@@ -74,6 +77,14 @@ jobs:
           key: ${{ runner.os }}-cargo-main-full-${{ hashFiles('**/Cargo.lock') }}
           restore-keys: |
             ${{ runner.os }}-cargo-main-full-
+
+      - name: Setup Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: "22"
+
+      - name: Policy JS unit tests
+        run: npm run test:policies
 
       - name: cargo test (non-PG)
         run: cargo test --all-targets -- --skip _pg --skip pg_ --skip postgres
